% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agconversions-package.R
\docType{package}
\name{agconversions}
\alias{agconversions}
\alias{agconversions-package}
\title{Convenient Imperial to Metric Conversion of Agriculture Statistics}
\description{
Are you tired of having to use calculators to convert between acres, hectares, bushels etc.?
Figuring out which data to use can get confusing when looking at large agriculture statistics
datasets with multiple variables including temperature, area, weight, and application rates.
}
\details{
Use these functions to call conversion ratios to your ag statistics. These functions convert some
of the most used imperial measurements in agriculture
to metric/SI: acres to hectares, acres to sq meters, pounds to metric tonnes,
farenheit to celcius and pounds per acre to kilograms per hectare.
}
\note{
This package isn't actually meant to be serious. It's just for
learning purposes and for completing an assignment for submission.
}
\seealso{
\code{\link{hectares}}, \code{\link{mtonnes}}, \code{\link{sqmeters}}, \code{\link{celcius}}, \code{\link{kgha}}
}
\author{
G Braich
}
