% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celcius.R
\name{celcius}
\alias{celcius}
\title{Converting farenheit to celcius}
\usage{
celcius(x, plot_it = FALSE)
}
\arguments{
\item{x}{The vector to be converted from degrees farenheit to degrees celcius}

\item{plot_it}{Display a plot of \code{x} vs the output? Use logical.
\code{FALSE} by default.}
}
\value{
A vector that is the:
\itemize{
     \item celcius (for \code{celcius})
}
 for degrees farenheit of \code{x}.
}
\description{
This function takes an imperial measurement for temperature, farenheit, and converts it into metric, specifically, celcius (not kelvin!)
}
\details{
If you really want to see more, check out all \href{http://ocw.tufts.edu/data/32/374423.pdf}{ag conversions data}
}
\examples{
celcius (32)
}
