% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgha.R
\name{kgha}
\alias{kgha}
\title{Converting application rate pounds per acre to kilograms per hectare}
\usage{
kgha(x, plot_it = FALSE)
}
\arguments{
\item{x}{The vector to be converted from pounds per acre to kilograms per hectare}

\item{plot_it}{Display a plot of \code{x} vs the output? Use logical.
\code{FALSE} by default.}
}
\value{
A vector that is the:
\itemize{
     \item kgha (for \code{kgha})
}
 for pounds per acre of \code{x}.
}
\description{
This function converts application rates commonly used for fertilizer or pesticide application.
}
\details{
If you really want to see more, check out all \href{http://ocw.tufts.edu/data/32/374423.pdf}{ag conversions data}
}
\examples{
kgha (100)
}
