% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtonnes.R
\name{mtonnes}
\alias{mtonnes}
\title{Converting pounds to metric tonnes}
\usage{
mtonnes(x, plot_it = FALSE)
}
\arguments{
\item{x}{The vector to be converted from pounds to metric tonnes}

\item{plot_it}{Display a plot of \code{x} vs the output? Use logical.
\code{FALSE} by default.}
}
\value{
A vector that is the:
\itemize{
     \item mtonnes (for \code{mtonnes})
}
 of pounds of \code{x}.
}
\description{
This function takes an imperial measurement for weight, pounds, and converts it into metric, specifically, metric tonnes
}
\details{
If you really want to see more, check out all \href{http://ocw.tufts.edu/data/32/374423.pdf}{ag conversions data}
}
\examples{
mtonnes(4700)
}
