% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{multi}
\alias{multi}
\title{Generic Multiplier Function}
\usage{
multi(x, a, plot_it)
}
\arguments{
\item{x}{Vector multiplied by some number.}

\item{a}{Number to multiply \code{x} by.}

\item{plot_it}{Display a plot of \code{x} vs the output? Use logical.}
}
\value{
The vector \code{x}, multiplied by \code{a}.
}
\description{
This function multiplies any vector. I'm making it internal,
by not \code{export}ing, because I don't want users to use it.
Why not? Because I said so!
}
