% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqmeters.R
\name{sqmeters}
\alias{sqmeters}
\title{Converting acres to square meters}
\usage{
sqmeters(x, plot_it = FALSE)
}
\arguments{
\item{x}{The vector to be converted from acres to square meters}

\item{plot_it}{Display a plot of \code{x} vs the output? Use logical.
\code{FALSE} by default.}
}
\value{
A vector that is the:
\itemize{
     \item sqmeters (for \code{sqmeters})
}
 of acres of \code{x}.
}
\description{
This function takes an imperial measurement for land, acres, and converts it into metric, specifically, meters squared
}
\details{
If you really want to see more, check out all \href{http://ocw.tufts.edu/data/32/374423.pdf}{ag conversions data}
}
\examples{
sqmeters(20000)

}
